[iOS][media-permissions] Custom alert UI 1/3
This CL updates the interface for AlertRequest and AlertResponse to use
a double array of button configs instead of a single one. A follow-up CL
will update the consumer to use the double vector to show the buttons in
the horizontal and vertical order arranged by the double vector.

Also, since the strings for permissions overlay has been added, this CL
also updates the strings used in the permissions overlay dialog as well
as the button configs, and adds unit tests for it.

Bug: 1356768
Change-Id: I60fb85163a7ea3d71aabe9f23b2843dccdd12f83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4112057
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Reviewed-by: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1089118}