media/gpu/v4l2: Match video and media devices by bus_info first
The bus_info field in the V4L2 media controller device [1] and video
device capability [2] provide a better, more specific matching between
devices than the driver field. The driver field can cause confusion if
there are multiple instances of hardware supported by the same driver.
One such case can be found on the iMX.8 SoCs, which has two Hantro
decoder blocks, a G1 and a G2.

Try to match by the bus_info fields if they are filled in in both
data structures, and fall back to using the driver fields if one or
both of them are empty. This workaround is for older mtk-vcodec drivers
which didn't fill in the bus_info field on the media device.

Note that this is still not the officially supported way of matching
devices. The official way is to use the MEDIA_IOC_ENUM_ENTITIES ioctl
to find a match for the video device node's major and minor numbers.

[1] Retrieved from /dev/mediaN with MEDIA_IOC_DEVICE_INFO ioctl
[2] Retrieved from /dev/videoN with VIDIOC_QUERYCAP ioctl

Bug: b:258331312
TEST=Run vdaTests on iMX.8 system and selects correct pairing media device
     Needs kernel patch for hantro driver to not use static bus_info
     value

Change-Id: I776ebebd1712dc4b329b83d408c685beeabb9bfc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4113217
Reviewed-by: Steve Cho <stevecho@chromium.org>
Auto-Submit: Chen-Yu Tsai <wenst@chromium.org>
Commit-Queue: Chen-Yu Tsai <wenst@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1089136}