cc::PaintImage::Decode: Consolidate arguments
The function cc::PaintImage::Decode takes as parameters an SkImageInfo,
a pointer, and an SkColorSpace. Merge the SkColorSpace into the
SkImageInfo, and merge the SkImageInfo and pointer into an SkPixmap.

Propagate this to DecodeFromGenerator and DecodeFromSkImage.

Update the callers in GpuImageDecodeCache and
SoftwareImageDecodeCacheUtils.

The GpuImageDecodeCache helper function DrawAndScaleImage is very
difficult to follow because the RGB and YUV paths are almost completely
disjoint, but are threaded through a single function. Separate these
into two functions and clean them up.

Bug: 1404000
Change-Id: Iddff08d84d2660913a73b9ed61e618ad7f70b625
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4128197
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1089156}