Remove kSkipAccessibilityPaintChecks from ArcNotificationContentView
The paint checks were failing due to ArcNotificationContentView having
an empty name. That was fixed in crrev.com/c/3883828. There do not
appear to be any other tests failing, so it should be safe to stop
skipping the accessibility paint checks on this view.

Bug: 1348081, 1218186
Change-Id: Ia6e199014e4dff499e3830331c5d1ce3760aae76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3956874
Reviewed-by: Toshiki Kikuchi <toshikikikuchi@chromium.org>
Commit-Queue: Joanmarie Diggs <jdiggs@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1089174}