Remove FocusBehavior; enable paint checks on AssistantWebContainerView
The a11y paint checks failed because AssistantWebContainerView
had FocusBehavior::ALWAYS (as long as there were contents), but lacked
both an accessible name and a valid accessible role. The expected fix
would be to give this view an accessible name and role so that screen
readers have something to present when the view gains focus. However,
it appears that this view's FocusBehavior might no longer be needed:

* crrev.com/c/1366757 added FocusBehavior::ALWAYS to AssistantWebView,
  along with handling that transferred focus to to the contents view.
  This suggests the AssistantWebView was not end-user focusable.
* crrev.com/c/1999042 removed that handling from AssistantWebView, but
  left the FocusBehavior::ALWAYS in place.
* crrev.com/c/2040199 removed AssistantWebView, moving its
  FocusBehavior::ALWAYS to AssistantWebContainerView

Because the AssistantWebContainerView lacks both a valid name and a
valid role, ChromeVox presumably couldn't present anything to the
user when it gained focus, something that would have been noticed
performing the tests described in crrev.com/c/1999042. That it wasn't
noticed also suggests this view is not end-user focusable.

Remove FocusBehavior::ALWAYS from AssistantWebContainerView. Having
done so, re-enable the accessibility paint checks which should now
pass.

Bug: 1348081, 1218186
Change-Id: I9f68a1477ad30bcc3565b3bf50c731503b612912
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4136235
Reviewed-by: David Black <dmblack@google.com>
Commit-Queue: Joanmarie Diggs <jdiggs@igalia.com>
Reviewed-by: Tao Wu <wutao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1089175}