[ReadingList] Delete account data upon signout
When a ReadingListModelImpl instance is explicitly used for account
data (sync the transport for reading lists), this data should be deleted
when the user signs out.

This logic can be achieved trivially in ReadingListSyncBridge, but an
enum needs to be plumbed to be able to trigger it only for the case
where the bridge is operating in transport mode (as the default case
when the user turned on sync shouldn't behave this way).

In order to implement deletion very reliably, including edge cases like
orphan data or untracked sync entities, the patch add a specific API
in ReadingListModelStorage that is capable of deleting all data at
once.

Change-Id: I35f93ab6a8d9f373149325859ffdfa9b0cf3e771
Fixed: 1402219
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4118747
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Brandon Wylie <wylieb@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1089181}