[ios] Scroll up when looking for "sign-in" button in settings
The "sign-in" button is at the top of the settings. Change the
code looking for the button to scroll up instead of down, thus
allowing to see a screenshot of the top of the settings (where
the button is supposed to be located) in case of failure.

This help debugging issues when the fake identity provider is
not correctly installed (i.e. when the tests uses the public
provider instead of the fake provider).

Bug: 1201182
Change-Id: I7dc503da72cf84e73ee8a07579b630d55d0da7fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4136377
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1089206}