[ios] Use an helper method in AuthenticationFlow
Use an helper method to deal with the result of checking whether the
identity is subject to "parental control". This simplifies the block
passed to the IsSubjectToParentalControls() method and reduces the
chances of inadvertently retaining "self" in the method.

This will also make the conversion to SystemIdentityManager simpler.

Bug: 1201182
Change-Id: I994970bc64f3edd75b3765f4b7454d8740e89269
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4134855
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Commit-Queue: Nohemi Fernandez <fernandex@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1089215}