Support multiple navigations in PageLoadMetricsWaiter
We use PageLoadMetricsTestWaiter to wait for the values of our pageload
metrics to be received in the PageLoadMetricsObserver in browser tests.

However, currently when we start a navigation, we create a
WaiterMetricsObserver which is a subclass of PageLoadMetricsObserver.
When we start another navigation from the current navigation, (except
the soft navigation/same document navigation), we finish the current
navigation, and we call OnCommit. In OnCommit, it calls
PageLoadMetricsTestWaiter::AddObserver, and checks if it has a
WaiterMetricsObserver, which is existed. So the DCHECK fails, we
can only use PageLoadMetricsTestWaiter for a single page load.

In order to make PageLoadMetricsTestWaiter work with multiple
navigations, in PageLoadMetricsTestWaiter::OnCommit, if we check
that if |did_add_observer_| is true, we should not call AddObserver
to add a new WaiterMetricsObserver.

Additionally, it seems that attach_on_tracker_creation_ works very
similar with did_add_observer_, only in OnTrackerCreated, it creates a
observer earlier. So we keep it in OnTrackerCreated and will look into
if we can delete it completely in a following CL.

There are some tests which can be benefit from this change, I will
enable then in the following CLs.

Bug: 1393425, 1224780
Change-Id: Ie0c238e69a1cbbd4b987c09af1a02df59b6cf90b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3983589
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Commit-Queue: Lan Wei <lanwei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1089259}