PrefModelAssociator: Remove EnforceRegisteredTypeInStore()
This method checked, at pref registration time, that a pref's value in
the user pref store actually matches the registered type of the pref.
This was necessary in the past to support late pref registration, which
meant values could sometimes be written before a pref was registered
and thus before the correct type was known.
Late pref registration was removed years ago, so this justification no
longer holds. These days, it should not be possible for the persisted
value to have a wrong type type (and if that ever happened, e.g. due
to data corruption, it would be an issue in the preferences system
itself, not with Sync).

Bug: 1404937
Change-Id: I6c7e4e3102868775b96c3098f8ffa9337c4d3776
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4136728
Reviewed-by: Maksim Moskvitin <mmoskvitin@google.com>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1089270}