Upgrade the select file dialog bridge to UTTypes
This was previously committed with https://crrev.com/c/4133645 and
subsequently reverted. This version changes the test to match the
new code.

A side note: The documentation for `-[NSSavePanel
setAllowedFileTypes:]` claims that "If no extension is given by the
user, the first item in the allowedFileTypes will be used as the
extension for the save panel." This is not actually true; the
implementation of `-[NS(Local)SavePanel _preferredFileNameExtension]`
shows us that they (in 10.13 if not earlier) roundtrip through UTTypes
to use the preferred extension for the type instead. Therefore, the
change to content types doesn't actually change the end-user-visible
behavior, even if it does change the test expectations.

Bug: 1393401
Change-Id: I5c26f06f1f32cd9638e11e90602d7b78a98eca7a
Cq-Include-Trybots: luci.chromium.try:mac_chromium_10.15_rel_ng
Cq-Do-Not-Cancel-Tryjobs: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4133394
Commit-Queue: Tom Burgin <bur@chromium.org>
Auto-Submit: Avi Drissman <avi@chromium.org>
Reviewed-by: Tom Burgin <bur@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1089313}