Privacy Sandbox Settings: Refactor the privacy sandbox allowed logic
This CL moves the check on incognito and restricted accounts into a
separate method, from the existing IsM1PrivacySandboxApiEnabled which
takes into account also a |pref_name|. The requirement for this move is
because the vanilla Shared storage access check doesn't need to depend
on an any pref.

This CL also incorporates some changes made in crrev/c/4111347 like
the support for Status.

Bug: 1378703, 1395437
Change-Id: I7841fe84997d05d9d290c60f7618f395392e4876
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4134078
Reviewed-by: Olesia Marukhno <olesiamarukhno@google.com>
Commit-Queue: Rohit Agarwal <roagarwal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1089323}