shortcuts: Update UI to match spec for accelerator rows, and change lock icon logic
This CL updates the styling of shortcut rows to match the spec
(http://shortn/_sQ3wyVPpkJ). Specifically, it adds additional padding,
adds a line separator between shortcuts with multiple entries, and
aligns the shortcut description to the top of the container.

Many of the changes in this CL revolve around the conversion of
the AcceleratorRow component to use CSS Grid:
https://screenshot.googleplex.com/8N6LicbLfSiEkJq
To facilitate that, I moved the lock icon + its logic from
AcceleratorRow into AcceleratorView. I also updated the tests to reflect
this change.

Also, I updated the lock icon logic to consider the `locked` property
of the accelerator-info, which it did not consider before.

Screenshots:
- Customization disabled, before:
https://screenshot.googleplex.com/4FbSv8u5y6Wz5gT
- Customization disabled, after:
https://screenshot.googleplex.com/AWrBH98EKhbCJud
- Customization enabled, before:
https://screenshot.googleplex.com/3oreeH6tY9BZNYj
- Customization enabled, after:
https://screenshot.googleplex.com/8WSJG55vyLuW3kg

Bug: b/216049298
Test: browser_tests --gtest_filter=ShortcutCustomizationApp*
Change-Id: I59e58455049eaccd93fd0e327926528cac633127
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4133945
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Commit-Queue: Camden Bickel <cambickel@google.com>
Cr-Commit-Position: refs/heads/main@{#1089326}