chrome.bookmarks: Migrate to Promise-based extension APIs
This is a reland of https://crrev.com/c/4114863.

The previous test directly reassigned the chrome.bookmarks.create method, omitting an optional callback which is used in production code. Upon migrating chrome.bookmarks to the Promise API, this implementation became untenable because the callback could no longer be omitted in tests. The previously-omitted code is now being executed, causing an assertion failure in edit_dialog_test.ts.

To solve this, a setDebouncerForTesting method has been added to
api_listener.ts, which is called in the failing test.

See diff
https://chromium-review.googlesource.com/c/chromium/src/+/4120645/1..10

Bug: 1368302
Change-Id: Ib62182ee95c697a66ffd11cc5ba648b2d2a760a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4120645
Reviewed-by: John Lee <johntlee@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Cole Horvitz <colehorvitz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1089344}