[iOS] Disable OverScrollActions when in Fullscreen Mode on iPhones
While a user is in fullscreen mode, the user can still do a "pull down"
gesture and trigger the overscroll menu which creates a clunky user
journey. This CL fixes this issue so that the overscroll menu is
disabled (similar to iPad) while in fullscreen mode for iPhones.
iPads currently don't have the overscroll menu functionality, so we
only focus on behavior on the iPhone.

Bug: 1404255
Change-Id: I88b345ffde8f1ee7d0b334ede5927040650a5fb4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4133414
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1089372}