Update link and product text colors in the about:credits page
This is a follow up to CL:4078268. Based on design input, it was
suggested to use #1967d2 for link color (i.e. google-blue-700)
and #202124 (i.e. google-grey-900) for the title colors.

Thus, update --interactive-color to --google-blue-700, and remove
--product-text-color in favor of falling back to --primary-color,
which is already set to #202124.

This maintains the required 4.5:1 text contrast ratios.

See the following link for a screenshot of these updates:
https://bugs.chromium.org/p/chromium/issues/detail?id=1394531#c21

Bug: 1394531
Change-Id: Id85a968b2ab65c5a49d16f41be5e1910d78f5636
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4137233
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1089406}