input_injector_wayland: Await keyboard capability
The host dynamically detects if the seat is capable of
keyboard or not. The detection happens when the user first
interacts with the keyboard. This CL injects a fake key
upon the input injector creation, which causes the keyboard
capability to be acquired and user can then continue to inject
their inputs normally. This avoids user first keystroke from
being lost.

Bug: chromium:1291247
Change-Id: I835e5287dc6a1350e9f63ef3e8f1f68b7b59dad9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4108169
Reviewed-by: Lambros Lambrou <lambroslambrou@chromium.org>
Auto-Submit: Salman Malik <salmanmalik@chromium.org>
Commit-Queue: Salman Malik <salmanmalik@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1089418}