ash: VcEffectsDelegate::GetEffectState() returns absl::optional<int>
Make the return value of VcEffectsDelegate::GetEffectState() an
optional int. This is a cleaner solution for cases where the current
state of a hosted effect cannot be obtained e.g. the delegate is in a
bad state or encounters an error.

This fixes a case where the background blur selector was disappearing
from the VC bubble, because it's state of
"off" (CameraEffectsController::BackgroundBlurEffectState::kOff) had
the same value (-1) as VcEffectState::kUnusedId.

Bug: b/264470750
Change-Id: I3ead33518c5ad79f1b891bed074c9d724163244f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4135196
Commit-Queue: Roger Tinkoff <rtinkoff@google.com>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1089466}