[Tab Groups] Fix GroupIdToGroupIndexMap on addTab
If a tab is added that is not grouped the current TabGroupModelFilter
implementation assumes that group will be at the end of the TabModel
and sets that group's index to be at the end of the GroupIndexMap.

This is not always the case and as such it is possible to end up with a
situation where the order of tabs in the TabGroupModelFilter disagrees
with the order in TabModel. This results in the GTS behaving incorrectly
when moving or grouping tabs.

Fix this by updating the index dynamically whenever a tab is added that
is not during a reset.

Skip flaky test detection due to Filter initialization error.

Validate-Test-Flakiness: skip
Bug: 1382771
Change-Id: I7daaf194f0552055e4720ae8235a6757df9aa608
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4133994
Reviewed-by: Yue Zhang <yuezhanggg@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1089476}