[Web Payment] Remove PaymentInstruments from delegations test
Before this patch, a Java integration test for
paymentManager.enableDelegations() method installed service worker based
payment handlers using the PaymentInstruments API, which is planned to
be removed.

This patch removes the Java integration test for enableDelegations()
method and instead replaces it with cross-platform browsertests that run
on both Android and desktop platforms. The patch adds a cross-platform
way to check whether shipping or contact info section is displayed to
verify that enableDelegations() is working correctly to hide those
sections.

After this patch, the enableDelegations() tests are not using the
PaymentInstruments API.

Bug: 1327265
Change-Id: I1557b71875a02f369f031eb47269848695593dfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4135775
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1089477}