[Fuchsia] Destroy surface view in FlatlandWindow when the view is detached
In WebEngine we need to ensure that when a View that shows a web.Frame
is reattached to the View tree it displays only current content of the
page and not the old content that was rendered before the page was
detached. This CL duplicates the approach that was previously
implemented in ScenicWindow: the surface viewport is dropped whenever
the view is detached from the display. Also, delegate events are emitted
to ensure that FlatlandSurface is re-initialized in the GPU process.

Bug: 1371497
Change-Id: I55c69f9e0f81f95e28065df65e42ce66180d4e4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4116092
Reviewed-by: Emircan Uysaler <emircan@chromium.org>
Commit-Queue: Sergey Ulanov <sergeyu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1089490}