Use actual scroll corner size as the minimum resize area
This patch replaces a fixed minimum resize size of 15 with the actual
size of the resize corner. This fixes an issue where custom scroll
corner sizes were not accounted for, as well as an issue where zoom
was not included.

This does not fix crbug.com/1403486 but simplifies fixing that issue.

Bug: 1403486
Change-Id: I76ec4452f8196fc33590663672b3c82cc59ae0f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4137577
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1089500}