Replace check_call with run() for llvm-profdata merge command
subprocess module per python docs recommend using run() to capture
stdout and stderr as noted here. updating the call to run(), and
adding logging for stdout to ensure that we're capturing all the
info provided. also distinguishing stderr and output for logging
until the point of err.

this also changes the _call_profdata_tool() signature. It never
returned invalid profiles, and returned an empty array on success
which was never utilized. this change cleans this up.

Bug: 1401044
Change-Id: I2d5273ac2adbb33dc6fd8df29f2d0927535dfc77
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4112286
Reviewed-by: Zhaoyang Li <zhaoyangli@chromium.org>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Cr-Commit-Position: refs/heads/main@{#1089528}