Add EXPECT_SKCOLORS_CLOSE macro
This CL moves gfx::ColorsClose from image_unittest_util.cc to
sk_color_eq.cc and wraps it in a macro. This allows for tests to
compare SkColors for closeness and get an assertion message that
includes the RGBA value of the color as well as the expressions used in
the comparison.

Change-Id: I91bcf2e4f48458950dfe5e070314a6cd79bb37bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4038344
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Michael Tang <tangm@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1089569}