gpu: win: Change opaque root surface SkColorType from kRGBA_8888 to kRGB_888x
This CL makes it so when Viz requests an opaque color type for the root
surface, we'll have the SkiaOutputDevice advertise that it Skia should
draw into it with an opaque buffer format. We already convert the
SkAlphaType to the correct DXGI_ALPHA_MODE for the DComp surface and
DXGI swap chain. This change may speed up some blending operations
internal to Skia.

Bug: 1378741
Change-Id: Iacfda842f2d57ebc44b099ba60a42a1277b6b101
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4052722
Commit-Queue: Michael Tang <tangm@microsoft.com>
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1089588}