[APS] Allow APS installed OEM apps to perform updates.
This CL updates apps with an install source of kOem to perform manifest
updates without prompting the user. This is needed because the
information used to install the app could be outdated, and it's bad
user experience for the user to see the prompt for an app that has
just been installed.

Fixed: b/262465858
Change-Id: I6b5de0544a37d92fff76193f8e86528bb8d7f483
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4135363
Commit-Queue: Alan Cutter <alancutter@chromium.org>
Auto-Submit: Jeevan Shikaram <jshikaram@chromium.org>
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1089612}