Files app: Skip sub-directory enumeration when necessary
Some volumes (eg. network filesystems) have the delayExpansion flag set
in their DirectoryTree. When expanding a directory (eg. via double-click
in the file list or directory tree), this flag is meant to prevent a
recursive walk of that directory's children.

This works when navigating via the file list, but if the directory is
expanded via the directory tree, a new SubDirectoryItem is created for
each child of the expanded directory, and it did not fully account for
its parent's delayExpansion flag: it would still enumerate at least
one element of its contents to determine whether the expansion icon
should be shown. This behaviour is still expensive in terms of network
calls.

Fix this behaviour so that when expanding directories via the directory
tree, we don't enumerate the contents of their immediate sub-directories
if the delayExpansion flag is true.

This CL also narrows the scope of which volumes delay expansion (from
all network volumes, which unnecessarily included DriveFS, to just SMB
and FSP volumes), and adds UMAs on how long it takes (per root type)
to expand a directory via the expansion icon.

Bug: b:255701094
Test: browser_tests --gtest_filter="*directoryTreeExpandFolder*"
Change-Id: I484f0b370450434a038a449b91381d8ed903f3d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4113091
Reviewed-by: Josh Simmons <simmonsjosh@google.com>
Reviewed-by: Nigel Tao <nigeltao@chromium.org>
Reviewed-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Josh Simmons <simmonsjosh@google.com>
Cr-Commit-Position: refs/heads/main@{#1089622}