Provide valid structure for input to EvalPolicy
This fuzzer uses an overlay of a structure that mirrors the real
classes however it did not correctly indirect the pointers to
pointers that were necessary to provide valid input. The overlaid
classes also now use raw_ptr so the fake classes should to.

Invalid argument types are also not provided as they simply hit:

[0104/164321.676:ERROR:sandbox_policy_base.cc(653)] Check failed: false.

See bug for an overview of how the structures should be laid out.

Bug: 1404590
Change-Id: I1cb96b47760e3c38998f648aac2e47a9fece5823
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4137217
Commit-Queue: Alex Gough <ajgo@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1089624}