[//media] Make APIs that generate legacy mailboxes private with friends
We are trying to eliminate legacy mailboxes. One of the remaining uses
is wrapper APIs in //media for video decode acceleration. As the
elimination of these APIs will stretch over the first half of 2023
while we roll out NaCL changes, this CL makes these APIs private with
friending of existing callsites to ensure that no new usages creep in.
NOTE: This necessitates moving CommandBufferHelperImpl and
PictureBufferManagerImpl out of anonymous namespaces so that they can
be referenced for friending.

Bug: 1382031
Change-Id: I9def86d36321e04896bba4dafbbba561226d8cdf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4134852
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1089663}