[ios] Cleanup signin provider test constants
Add a target to export the accessibility identifiers of UI elements
exposed for EarlGrey tests (this is only used by fake UI elements,
so the target is marked as `testonly=true`).

Inline GetManagedEmailSuffixes() since it is only used in a single
location and this makes the code easier to read and avoid having to
convert the code when switching from the old to the new provider.

Remove unused imports and dependencies.

Bug: 1201182
Change-Id: Idbe6d04aec4c537b71d85b30d935490ecd711f8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4134858
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1089692}